+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.15.2] - 2020-08-17
+
 ### Fixed
 
 - Add calls to simulation for product sellers on compatibility layer.